Extract the shared amount parameter into a reusable payments type
Move the `AmountParameter` request type out of `z_send_many` into
`json_rpc::payments` (with accessor methods) and extract the shared
`build_request` amount-parsing helper, so the type and parsing can be reused by
other send-path RPC methods. Update the method trait and OpenRPC generation to
reference the relocated type, and keep the `AmountParameter` import in the
wallet-gated block so the merchant-terminal build still compiles.
