+++ zallet/src/components/json_rpc/methods/openrpc.rs
@@ -1,5 +1,6 @@
 use std::borrow::Cow;
 
+use crate::components::json_rpc::payments::AmountParameter;
 use documented::Documented;
 use jsonrpsee::core::{JsonValue, RpcResult};
 use schemars::{JsonSchema, Schema, SchemaGenerator, generate::SchemaSettings};
@@ -7,7 +8,7 @@ use serde::Serialize;
 
 // Imports to work around deficiencies in the build script.
 #[cfg(zallet_build = "wallet")]
-use super::{super::asyncop::OperationId, recover_accounts, z_send_many};
+use super::{super::asyncop::OperationId, recover_accounts};
 
 // See `generate_rpc_help()` in `build.rs` for how this is generated.
 include!(concat!(env!("OUT_DIR"), "/rpc_openrpc.rs"));
